+++ docs/release_notes.rst
@@ -4,7 +4,7 @@ Release Notes
 v4.1.0
 ------
 
-* Caught HA error for uploading objects and raised it as a `swift.UnavailableError`
+* Caught HA error for uploading objects and raised it as a `swift.FailedUploadError`
 
 v4.0.0
 ------